sc - move Todo class + tests into an entity folder, updated imports too
